src/lib/utils/__tests__/storage.test.ts
// TODO:: FIX ONCE FIREFOX STORAGE IS SOLVED
test('test', () => {
  expect(true).toBe(true)
})
// // Mock needs to be before any imports
// const mockStorage = {
//   local: {
//     get: jest.fn().mockResolvedValue({}),
//     set: jest.fn().mockResolvedValue(undefined),
//     remove: jest.fn().mockResolvedValue(undefined)
//   }
// }

// // Mock the module before importing
// jest.mock('webextension-polyfill', () => ({
//   storage: mockStorage
// }))

// // Now we can import our modules
// import { SecureStorage } from '../storage'


// describe('SecureStorage', () => {
//   const mockWallet = {
//     mnemonic: 'test mnemonic',
//     masterSeed: new Uint8Array([1, 2, 3]),
//     accounts: {},
//   }

//   const mockPassword = 'testPassword123!'

//   beforeEach(() => {
//     // Clear all mocks before each test
//     jest.clearAllMocks()
    
//     // Reset mock implementations
//     mockStorage.local.get.mockResolvedValue({})
//     mockStorage.local.set.mockResolvedValue(undefined)
//     mockStorage.local.remove.mockResolvedValue(undefined)
//   })

//   describe('encrypt', () => {
//     it('should encrypt data with password', async () => {
//       const encrypted = await SecureStorage.encrypt(mockWallet, mockPassword)

//       expect(encrypted).toHaveProperty('data')
//       expect(encrypted).toHaveProperty('salt')
//       expect(encrypted).toHaveProperty('iv')
//       expect(encrypted).toHaveProperty('version')
//       expect(encrypted).toHaveProperty('iterations')
//       expect(encrypted).toHaveProperty('timestamp')
//       expect(encrypted).toHaveProperty('attempts')
//     })

//     it('should generate different ciphertexts for same data', async () => {
//       const encrypted1 = await SecureStorage.encrypt(mockWallet, mockPassword)
//       const encrypted2 = await SecureStorage.encrypt(mockWallet, mockPassword)

//       expect(encrypted1.data).not.toBe(encrypted2.data)
//       expect(encrypted1.salt).not.toBe(encrypted2.salt)
//       expect(encrypted1.iv).not.toBe(encrypted2.iv)
//     })
//   })

//   describe('decrypt', () => {
//     it('should decrypt encrypted data correctly', async () => {
//       const encrypted = await SecureStorage.encrypt(mockWallet, mockPassword)
//       const decrypted = await SecureStorage.decrypt(encrypted, mockPassword)

//       expect(decrypted).toEqual(mockWallet)
//     })

//     it('should fail with wrong password', async () => {
//       const encrypted = await SecureStorage.encrypt(mockWallet, mockPassword)
//       await expect(
//         SecureStorage.decrypt(encrypted, 'wrongPassword')
//       ).rejects.toThrow('Invalid password')
//     })

//     it('should handle version mismatch', async () => {
//       const encrypted = await SecureStorage.encrypt(mockWallet, mockPassword)
//       encrypted.version = 999

//       await expect(
//         SecureStorage.decrypt(encrypted, mockPassword)
//       ).rejects.toThrow('Incompatible wallet version')
//     })

//     it('should handle too many attempts', async () => {
//       const encrypted = await SecureStorage.encrypt(mockWallet, mockPassword)
//       encrypted.attempts = 3
//       encrypted.timestamp = Date.now()

//       await expect(
//         SecureStorage.decrypt(encrypted, mockPassword)
//       ).rejects.toThrow('Wallet is locked')
//     })
//   })

//   describe('storage operations', () => {
//     it('should store and retrieve wallet', async () => {
//       await SecureStorage.saveWallet(mockWallet, mockPassword)
      
//       const storedWallet = await SecureStorage.loadWallet(mockPassword)
//       expect(storedWallet).toEqual(mockWallet)
      
//       expect(mockStorage.local.set).toHaveBeenCalled()
//       expect(mockStorage.local.get).toHaveBeenCalled()
//     })

//     it('should handle storage errors', async () => {
//       mockStorage.local.set.mockRejectedValue(new Error('Storage error'))
      
//       await expect(
//         SecureStorage.saveWallet(mockWallet, mockPassword)
//       ).rejects.toThrow('Storage error')
//     })

//     it('should handle missing wallet', async () => {
//       mockStorage.local.get.mockResolvedValue({})
      
//       await expect(
//         SecureStorage.loadWallet(mockPassword)
//       ).rejects.toThrow('No wallet found')
//     })
//   })

//   describe('password attempts', () => {
//     it('should track failed attempts', async () => {
//       const encrypted = await SecureStorage.encrypt(mockWallet, mockPassword)
      
//       await expect(
//         SecureStorage.decrypt(encrypted, 'wrong1')
//       ).rejects.toThrow()
      
//       await expect(
//         SecureStorage.decrypt(encrypted, 'wrong2')
//       ).rejects.toThrow()
      
//       expect(encrypted.attempts).toBe(2)
//     })

//     it('should lock after max attempts', async () => {
//       const encrypted = await SecureStorage.encrypt(mockWallet, mockPassword)
      
//       // Simulate max attempts
//       for (let i = 0; i < 5; i++) {
//         try {
//           await SecureStorage.decrypt(encrypted, 'wrong')
//         } catch (e) {
//           // Expected
//         }
//       }
      
//       await expect(
//         SecureStorage.decrypt(encrypted, mockPassword)
//       ).rejects.toThrow('Too many failed attempts')
//     })
//   })
// }) 